// src/navigation/types.ts
export type RootStackParamList = {
  Loader: undefined;
  Onboard: undefined;
  MainTabs: undefined;
  Math: undefined; 
  MathPlay: { level?: number } | undefined; 
  Hunt: undefined; 
  HuntPlay: { level?: number } | undefined; 
};

export type MainTabParamList = {
  Math: undefined;
  Hunt: undefined;
  Room: undefined;
  Facts: undefined;
  Tales: undefined;
};